requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg[binary]>=3.1
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1
gunicorn==22.0.0
